Fix NPE when empty URL in organization module
One of the feature of BLTI B2 module is to pull in a feed or URL to be
displayed in the module. This feature can be disabled by not putting
any URL in the module configuration page. Unfortunately, this feature
has a bug when you don't put any URL. It generates a NPE caused by an
uninitialized variable.