Fix an URL manipulation vulnerability
create.jsp and modify.jsp now check for the correct entitlement
(course.content.CREATE / course.content.MODIFY), but there is still an
insecure direct object reference to the content pk1 in the URL. See
below for example of how I can exploit this:
1. Enrol a user in two courses: one as student, one as TA
2. As this user, access course as student and note down content pk1s for
an existing content folder and LTI link
3. Access the second course as TA. From a content folder, Tools > More
Tools > Basic LTI Tools. In the URL, change the content pk1 to that of
an existing content folder in another course as noted in step 2. Reload
the page. Complete the creation wizard and submit. Notice you have added
content to the first course despite the fact you only have appropriate
entitlements in the second course.
4. Access the second course as a TA and “edit” an existing Basic LTI
link. In the URL, change the content pk1 to that of an existing LTI link
in another course as noted in step 2. Reload the page. Make changes.
Submit. Notice your changes go through even though you receive an error.